+++ setRec.h
@@ -21,11 +21,11 @@ int sr_mark(const SR_Base *, const unsigned long *, size_t,
         char);
 
 // Output Sets with Particular Mark Status
-long long sr_query(const SR_Base *, char, char,
+ssize_t sr_query(const SR_Base *, char, char,
         void (*)(const unsigned long *, size_t));
 
 // Output Sets with Particular Mark Status, for Parallelism
-long long sr_query_parallel(const SR_Base *, char, char,
+ssize_t sr_query_parallel(const SR_Base *, char, char,
         size_t, size_t, void (*)(const unsigned long *, size_t));
